Modified data api: rename project and app name to avoid conflicts, fixed mysqlclient error
