+++ mbtb_app/resources/apis/data/config/secret_key.cnf
@@ -1,0 +1,1 @@
+y^+fox(&)g1de5(#kj1ln^5l(isbcqom+&hdfh8hyiz_hmit8d
