+++ mbtb_app/resources/apis/data/data/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for data_app project.
+Django settings for data project.
 
 Generated by 'django-admin startproject' using Django 2.2.6.
 
@@ -20,12 +20,13 @@ BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 # See https://docs.djangoproject.com/en/2.2/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = 'y^+fox(&)g1de5(#kj1ln^5l(isbcqom+&hdfh8hyiz_hmit8d'
+with open('config/secret_key.cnf') as sk:
+    SECRET_KEY = sk.read().strip()
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
 
-ALLOWED_HOSTS = []
+ALLOWED_HOSTS = ['*']
 
 
 # Application definition
@@ -38,7 +39,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'rest_framework',
-    'dataapp',
+    'mbtb',
 ]
 
 MIDDLEWARE = [
@@ -51,7 +52,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'data_app.urls'
+ROOT_URLCONF = 'data.urls'
 
 TEMPLATES = [
     {
@@ -69,7 +70,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'data_app.wsgi.application'
+WSGI_APPLICATION = 'data.wsgi.application'
 
 
 # Database
@@ -78,11 +79,9 @@ WSGI_APPLICATION = 'data_app.wsgi.application'
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.mysql',
-        'NAME': 'mbtb_db',
-        'USER': 'root',
-        'PASSWORD': 'yinino0617',
-        'HOST': 'localhost',
-        'PORT': '3306',
+        'OPTIONS': {
+            'read_default_file': os.path.join(BASE_DIR, 'config/test_db.cnf'),
+        }
     }
 }
