+++ mbtb_app/resources/apis/data/data/urls.py
@@ -1,4 +1,4 @@
-"""data_app URL Configuration
+"""data URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/2.2/topics/http/urls/
@@ -15,7 +15,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path
-from dataapp import views
+from mbtb import views
 
 from django.http import HttpResponse
