+++ mbtb_app/resources/apis/data/mbtb/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class DataappConfig(AppConfig):
-    name = 'dataapp'
+    name = 'mbtb'
