--- mbtb_app/resources/apis/data_app/data_app/__init__.py
@@ -1,2 +1,0 @@
-import pymysql
-pymysql.install_as_MySQLdb()
